Avoid zombie child processes
We have two processes that `Command::new` but do not take care of waiting
for child processes. We can either call wait or let the kernel know that
we don't care by ignoring `SIGCHLD`. If we don't do this, any child
process will turn into a zombie that will consume resources and litter
the process tree.

The solution uses `nix::sys::signal` unconditionally, since leftwm is
an X11 wm, and it's reasonable to assume a suitable `nix::sys`
environment.