Fix subcommands continuing in background, fix clap errors showing for valid arguments to subcommands (#378)
* Wait for subcommand to exit

We were letting subcommands continue to run in the background. Now we
wait for them to exit before quitting. That way we don't lose input
control over the subcommand.

* Get subcommand arguments working

* Extract subcommands and descriptions into a single map

That way new subcommands (or changes to existing ones) need only happen
in one place, which will be less error prone.

* Add self-documenting variable for clarity

* Update and improve documentation

* Change short --version flag from -v to -V

This matches the other commands in the leftwm program suite.