Add GoToPreviousTag feature to unconditionally go to previous tag
Similarly to GoToTag feature, I wanted a way to unconditionally go to
previous tag, without having to press the key for the current tag.

No default keybindings have been assigned for this feature.

I did not find an up to date list of available commands so I did not
know where to document this new command.