test(core): reproduce issue with default_layout
These tests attempt to reproduce the issues with PR #1301 encountered by @/grndplane.

Specifically, needed to extend the MockDisplayServer with the whole "inititial events" stuff the X11 display servers do on initialization to create workspaces for the available screens.

Once that was done, I wrote tests to illustrate how default_layout is handled by the Manager when the `output` of a configured workspace doesn't match any of the available screens (as well as what it does when they do match, and when only some of them match).

Through these tests, I was able to determine that the issue grndplane encountered was most likely due to a misconfiguration of the workspaces `output` fields as opposed to there being a bug in my code.