fix sloppy focus after window closes
After a window closes, the focus is updated to another window. For
sloppy focus this happens by queueing FocusWindowUnderCursor. However,
when this action is being executed, there is a check for change in focus
which happens _after_ the window being destroyed has been removed from
the window history. This in turn causes us to think there is nothing to
do and we exit early.

Push a None focus onto the window history to force an update. After all,
we don't actually have anything in focus at the moment.

fixes #1295