fix sloppy focus with margins
Commit 4f765b4ba14 ('Various fixes') claims to 'make it work for
non-sloppy focus' but in doing so, it breaks it for sloppy focus by
forcing the focus to the root window. This doesn't seem to be the
correct place to make this kind of decision anyway.

Removing this bit of code also doesn't really seem to break anything
obvious on ClickTo or Driven focus, so we might as well get rid of it.

fixes #1108