xlib, x11rb: fix ClickTo focus after switching tags
When switching tags back and forth in ClickTo focus, we can't click to
focus any windows. Only after the windows have come into focus through
other means, are we able to click on them.

So when switching tags, windows are hidden using the OnlyMove strategy.
In this case these windows are left in the `Normal` window state, even
if they aren't visible. When we switch back to the original tag, we then
don't call `toggle_window_visibility`, because we think there's nothing
to do. This in turn means we don't call `grab_mouse_events` on these
windows, so we're never receiving any click events that enable us to
click to focus.

So when marking windows invisible, always make sure we take them out of
the 'normal' state.