fix creating a scratchpad window on x11rb
On x11rb, scratchpad windows are handled like any normal window: the
window doesn't float in the designated area, and is just added to
the active tag's window stack. This is because the WindowCreate event
doesn't receive a usable pid and due to that, the window can't be
matched up to any of the active scratchpads.

This all happens because `xwrap::get_property` calls
`xproto::get_property` with `0` as the last argument, which effectively
tells it to return no items.

So set it to `1` instead.

fixes #1294