x11rb: allow legacy name to not exist
When opening a dialog in a Qt application like wireshark, the dialog
isn't mapped due to the fact that the window doesn't have a legacy name.
In the case of wireshark, this blocks further use of the application.

We store the legacy name in an `Option`, so we should be able to deal
with it being `None`. Translate any errors when fetching the legacy
name into it just not being available, pretty much like the xlib backend
does.

Fixes #1317