x11rb: getters: fix getting the output name
Getting the output name always fails, as indicated by the `FIX`
note, leading to leftwm-state to always say the current output is
`"output_name"`.

This happens because we're assuming we're still doing ffi here and as
such are trying to read a null-terminated string from some bytes. In
fact the data isn't null-terminated, at least not here. And since there
is no null-termination, the conversion always fails.

To get this to work, we can simply convert the data to `str` and drop
the whole CStr use.