feat(watchdog): deprecate leftwm-watchdog bin; gate leftwm behind feat
Previously in #1192, we broke the `leftwm` binary (which is the `leftwm-watchdog`
binary) out from the `leftwm` crate into the `leftwm-watchdog` crate to
improve incremental compile times. However, because the interface for
the `leftwm-watchdog` crate updates the installation command to
`cargo install leftwm leftwm-watchdog`, which is relatively unituitive,
in this commit, we revert #1192 to bring the `leftwm` binary
back into the `leftwm` crate to allow installing all of `leftwm` with
`cargo install leftwm`; this is nicer for "first time" users as it
reduces confusion. The `leftwm-watchdog` binary is actually the `leftwm`
binary, but that isn't always immediately obvious.

Any packagers who currently vendor `leftwm-watchdog` will need to remove
the `leftwm-watchdog` crate and instead ensure the `leftwm-watchdog`
feature is included in the build. Because this is a default feature,
most packages should be fine. However, any package using
`--no-default-features` will not vendor the `leftwm` binary unless the
feature `leftwm-watchdog` is explicitly added back, e.g.:
```
cargo build --release --no-default-features --features leftwm-watchdog
```

For developers, the way to save time when doing incremental compilations
is now to explicitly exclude the default feature `leftwm-watchdog`,
e.g.:
```
cargo build --release --no-default-features
```

This reverts commit 5293219827bf0d2a982bf548b5fdbb0c4cc7d235.

This closes #1323 and closes #1309.