feat!(config): Remove TOML support (implement #875) BREAKING CHANGELOG
Previously, as part of 0.4.0, we deprecated TOML so we could use RON.
This change removes TOML support entirely, relying on external binary
[`leftwm-config`](https://github.com/leftwm/leftwm-config/tree/main) to
handle migrations going forward. Any users who have not migrated to RON
from TOML will need to do so using `leftwm-config` or manual updates.

We will need to check that `leftwm-theme` and `leftwm-config` are
prepared to take on this change as well.

Fixes: #875